agent/requirements.txt
python-dotenv>=1.0.0
pydantic==2.11.7
httpx>=0.25.0
aiohttp>=3.8.0
mcp>=1.17.0
pydantic-ai>=0.0.13
google-adk>=1.13.0,<2.0.0
a2a-sdk>=0.3.4,<0.4.0
litellm>=1.76.3,<2.0.0
openinference-instrumentation-google-adk>=0.1.4,<0.2.0
arize-phoenix-otel>=0.13.0,<0.14.0
uvicorn>=0.30.0